### 2020-08-09 #2560 @whedon
[PRE REVIEW]: zalpha: an R package for the identification of regions of the genome under selection
**Submitting author:** @chorscroft (<a href="http://orcid.org/0000-0001-5679-5912">Clare Horscroft</a>)
**Repository:** <a href="https://github.com/chorscroft/zalpha" target ="_blank">https://github.com/chorscroft/zalpha</a>
**Version:** v0.2.0
