### 2020-09-04 #2638 @whedon
[REVIEW]: zalpha: an R package for the identification of regions of the genome under selection
**Submitting author:** @chorscroft (<a href="http://orcid.org/0000-0001-5679-5912">Clare Horscroft</a>)
**Repository:** <a href="https://github.com/chorscroft/zalpha" target ="_blank">https://github.com/chorscroft/zalpha</a>
**Version:** v0.2.0

@juanvillada , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
