### 2020-09-04 #2560 @lpantano
Hey @alachins, @FerRacimo would you like to review this tool? link to the guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
