### 2020-09-08 138ce99
Chnage to L_plus_R so that it will return zero rather than NA when n<2 in choose(n,2)
