### 2020-09-08 8bb8fce
Added definition of L and R to Zalpha_all documentation
