### 2020-09-08 fbb6ba2
Moved calculation of LR and L_plus_R so they are always calculated. Edit to final warning to not reference LR
