### 2020-09-08 eae5989
New developments to functions equal_vector, L_plus_R and Zalpha_all
