+++ NEWS.md
@@ -1,5 +1,9 @@
 # zalpha 0.2.0.9999 (Development version)
 
+* Improved the speed by editing the equal_vector function
+* L_plus_R now returns 0 instead of NA when n<2 for choose(n,2) 
+* Zalpha_all now always calculates the diversity statistics LR and L_plus_R
+
 # zalpha 0.2.0
 
 * Can now accept data with missing values
