### 2020-09-09 #2638 @juanvillada
Hi @lpantano
Hi @chorscroft

The PDF failed to compile. I will resume my review once it has been fixed.

Some notes for now:

1. The Repository URL is linking to the paper URL (https://github.com/chorscroft/zalpha/tree/master/paper) and not the repository itself.

2. It looks like the LICENSE is not one of the [OSI approved](https://opensource.org/licenses/alphabetical) software licenses.
