### 2020-09-09 #2638 @chorscroft
How odd, it compiled on the original thread for danielskatz on Aug 9th and I haven't changed anything! https://github.com/openjournals/joss-reviews/issues/2560

The licence is an MIT licence, I thought that was okay? Please let me know if it's not!
https://github.com/chorscroft/zalpha/blob/master/LICENSE.md
https://opensource.org/licenses/MIT

And apologies for linking to the wrong place, the submission form says "URL of the submitted software's git repository (containing the paper)" so I thought it was supposed to be a link to the paper in the software repository not the software itself, my fault. How can I fix it?

Thanks!
