### 2020-09-09 #2638 @juanvillada
> How odd, it compiled on the original thread for danielskatz on Aug 9th and I haven't changed anything! #2560

Excellent! The PDF of the paper looks good now. Thanks @danielskatz  ✅

> The licence is an MIT licence, I thought that was okay? Please let me know if it's not!
> https://github.com/chorscroft/zalpha/blob/master/LICENSE.md
> https://opensource.org/licenses/MIT

Yes, that one is ok. ✅
I just got confused with the other license file in the repo (https://github.com/chorscroft/zalpha/blob/master/LICENSE).

I will resume my review.
