### 2020-09-14 #2638 @lpantano
Hi @aursiber, would you like to review this paper? https://joss.readthedocs.io/en/latest/review_checklist.html
