### 2020-09-15 ddefe8a
Added code to force pairs of SNPs with a genetic distance greater than the biggest bin in the LD profile into the maximum bin
