### 2020-09-15 eddb955
Added test for when a pair of SNPs in a window have a genetic distance greater than the maximum bin in the LD profile
