### 2020-09-15 bbabd7a
Added documentation about max_dist
