### 2020-09-15 db62d3c
Added note about the changes to how max_bin is handled
