### 2020-09-22 #2638 @mdogucu
Thank you for considering me but I have never worked with genome data in my life. I think there is something unusual with review requests I get from JOSS and I am wondering how I can correct it. Either I have made a mistake in filing my areas of interest or JOSS is sending me random review requests. Is there any place I can update my reviewer settings for JOSS and edit my areas of expertise etc. ? I understand that finding reviewers is extremely difficult so I hope to ease the process for both the editors and myself.

Thank you @lpantano and everyone else who puts the time to run JOSS.
