### 2020-09-23 #2638 @aursiber
Sorry for the delay in responding. I have no competence in genomics to review this work.

> Hi @aursiber, would you like to review this paper? https://joss.readthedocs.io/en/latest/review_checklist.html
