### 2020-09-27 #2638 @juanvillada
The paper is well-written and references are appropriate. The authors make a good contribution to the field by building an open source software (R package) relevant to the identification of selective sweep in genomic regions. The package delivers what it aims. Overall, I think the paper is worth of publication.

Minor issues:

1. There are 4 co-authors listed on the paper while only one appears as the sole contributor of the software. I leave this consideration to the Editor (@lpantano) as I understand this is pretty normal in academic papers. Please see:
>Contribution and authorship: Has the submitting author (@chorscroft) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

2. The community guidelines are not explicit in the package repository. Please see:
>Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Thanks @chorscroft  and colleagues for making this OSS. 👍🏼   👍🏼   👍🏼
