### 2020-10-29 #2638 @chorscroft
> The paper is well-written and references are appropriate. The authors make a good contribution to the field by building an open source software (R package) relevant to the identification of selective sweep in genomic regions. The package delivers what it aims. Overall, I think the paper is worth of publication.
>
> Minor issues:
>
> 1. There are 4 co-authors listed on the paper while only one appears as the sole contributor of the software. I leave this consideration to the Editor (@lpantano) as I understand this is pretty normal in academic papers. Please see:
>
> > Contribution and authorship: Has the submitting author (@chorscroft) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> 1. The community guidelines are not explicit in the package repository. Please see:
>
> > Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> Thanks @chorscroft and colleagues for making this OSS. 👍🏼 👍🏼 👍🏼

Thank you @juanvillada for putting the time and effort to review my paper!

Apologies for the slow response, I wasn't sure if I was supposed to wait for all the reviews to be returned.

Response to issue 1:
Thanks for raising this point! I created the software and coded it all myself, however my co-authors have been invaluable to the project by:
- testing the code
- reading the documentation
- developing the statistics used in the code
- contributing knowledge about the science and biology behind the statistics
- reading and editing the text of the paper
It is standard in my field to include these people as authors on the paper.

Response to issue 2:
Thank you very much for bringing this to my attention, you are correct that these things were not explicit. I have now included a "Community Guidelines" section in the README file with instructions on how to raise issues, report bugs and seek support, as well as how to contribute via pull requests.

Thank you to everyone who has been involved in reviewing and editing this paper so far.
