### 2020-10-30 #2638 @marcrr
Sorry for the delay in writing this review.

Overall, this is a nice software and a nice contribution. It was easy to install and run. I have a few points, which I will structure following the Review checklist.

An small point: it would be nice to provide the compiled PDF for easier access by readers.

**Installation instructions: Is there a clearly-stated list of dependencies**

It was easy to install the dependencies, but I did not find them listed in one place. Following the criteria of the journal, it might be good to list them clearly somewhere.

**Automated tests**

I was unable to run test_check("zalpha") nor to find any reference to it in the documentation. Maybe I missed something? Could @chorscroft clarify whether there is an automated test, and whether this is it please?

**State of the field: Do the authors describe how this software compares to other commonly-used packages?**

This is my largest concern, the paper does not present any comparison to other software. There are many R packages which compute LD in some way or another.
The same concern applies to **References**.
