### 2020-11-08 ce245c1
Added a section on dependencies
