### 2020-11-09 b402973
Add Travis badge to README
