+++ README.md
@@ -1,5 +1,7 @@
 # zalpha
 
+  [![Travis build status](https://travis-ci.com/chorscroft/zalpha.svg?branch=master)](https://travis-ci.com/chorscroft/zalpha)
+
 R package for the zalpha suite of selection statistics. This package allows the user to run a suite of selection statistics over genomic data. The aim is to identify regions of the genome which have experienced a selective sweep.
 
 ## Installing
