### 2020-11-09 cd6e8a5
Added a paragraph and references about other packages
