+++ paper/paper.bib
@@ -4,6 +4,15 @@
   	Title = {{A Primer of Molecular Population Genetics}},
   	Year = 2019
 }
+@Article{Gautier:2017,
+    author = {{Gautier}, M., and {Klassmann}, A. and {Vitalis}, R.},
+    title = {rehh2.0: a reimplementation of the R package rehh to detect positive selection from haplotype structure},
+    journal = {Molecular Ecology Resources},
+    year = {2017},
+    volume = {17},
+    number = {1},
+    pages = {78-90},
+}
 @article{Jacobs:2016,
   title="{Refining the Use of Linkage Disequilibrium as a Robust Signature of Selective Sweeps}",
   author={{Jacobs}, G.S., and {Sluckin}, T.J. and {Kivisild}, T.},
@@ -41,6 +50,15 @@
   number={1},
   pages={23--35},
   year={1974},
+}
+@Article{Pfeifer:2014,
+    title = {PopGenome: An Efficient Swiss Army Knife for Population Genomic Analyses in R},
+    journal = {Molecular Biology and Evolution},
+    year = {2014},
+    author = {{Pfeifer}, B., and {Wittelsbuerger}, U. and {Ramos-Onsins}, S.E. and {Lercher}, M.J.},
+    volume = {31},
+    pages = {1929-1936},
+    doi = {10.1093/molbev/msu136},
 }
 @incollection{Vitti:2013,
   author       = {{Vitti}, J.J., and {Grossman}, S.R. and {Sabeti}, P.C.},
