### 2020-11-09 #2638 @chorscroft
@marcrr, thank you very much for the time and effort you put into reviewing our paper!

_An small point: it would be nice to provide the compiled PDF for easier access by readers._

Thanks for this point - I have now uploaded a pdf version of the paper to the repository.

**Installation instructions: Is there a clearly-stated list of dependencies**
_It was easy to install the dependencies, but I did not find them listed in one place. Following the criteria of the journal, it might be good to list them clearly somewhere._

Thank you for raising this point. There actually aren’t any required dependencies for this package; however, I have written a section into the README file to state the minimum R version required and to list the optional packages so users can be aware in advance.

**Automated tests**
_I was unable to run test_check("zalpha") nor to find any reference to it in the documentation. Maybe I missed something? Could @chorscroft clarify whether there is an automated test, and whether this is it please?_

Thank you for this point. The package does contain a suite of tests but they were inaccessible to the user. I have now created a Travis-CI account and linked the github repository and there is a badge in the README file to show the status. Travis-CI checks all the test whenever there is a push to the github repository. I hope this addresses this point.

**State of the field: Do the authors describe how this software compares to other commonly-used packages?**
_This is my largest concern, the paper does not present any comparison to other software. There are many R packages which compute LD in some way or another.
The same concern applies to References._

Thank you for raising this issue. I have added a paragraph with references to other packages that use similar methods to perform selection scans.

Thanks again for your review, it was very helpful!
