### 2020-11-12 #2638 @marcrr
@chorscroft Thank you for your revisions. They reply to almost all my concerns.

I still find the comparison to other packages very light, even with the new paragraph. I am not sure what the standards are for this journal, so I will let @lpantano make the final call on this.

Best regards
Marc
