### 2020-11-16 #2638 @lpantano
@marcrr. @chorscroft, I am checking with the editorial board about the amount of information when comparing with other packages. @marcrr, if you know other packages that are not there, feel free to mention it here. Even if we decides we don't need a lot, it would be good to have the best list possible and maybe we just ask for a little more information for each package. Thanks
