### 2020-11-17 #2638 @marcrr
From my point of view, it's less a question of specific other packages, than of discussing the relative advantages of the new package in a bit more detail.
