### 2020-11-18 #2638 @lpantano
thanks @marcrr, I got update from the BoE, and it seems enough to mention the tools so the reader has an idea where to go to learn more about the field. Since there is a description of the method already, we agreed this is almost enough.
@chorscroft, do you think you could add a little more about the other packages, if they target a different context, or they have some limitations your package doesn't, or a limitation of your package that the others can do? I think a couple of more sentences about the limitations and context for each should be enough to accept this. Thanks.
