+++ paper/paper.md
@@ -92,7 +92,7 @@ There are many statistics included in the package for adjusting for expected $r^
 
 The output of the functions is in list format. The SNP positions and the values of the statistic(s) are stored in vectors of equal length in the list. Users can then identify outlying SNPs in their data that are candidate regions for selection.
 
-There are a few other R packages that can be used for selection scans, although none utilise the $Z_{\alpha}$ statistics described here. These include `PopGenome` [@Pfeifer:2014], which calcualtes Kelly's $Z_{nS}$ among other methods, and `rehh` [@Gautier:2017], which implements extended haplotype homozygosity (EHH) and related statistics.
+There are a few other R packages that can be used for single population selection scans, although none utilise the $Z_{\alpha}$ statistics described here. `PopGenome` has the advantage of being able to read in multiple standard genomic data formats and perform other analyses as well as selection scans [@Pfeifer:2014]. The scans it performs are the CL and CLR methods by @Nielsen:2005. It can also calculate Kelly’s $Z_{nS}$ [@Kelly:1997] and Rozas’ ZA/ZZ [@Rozas:2001] for LD, which are similar to the basic $Z_{\alpha}$ statistic; however, neither use an LD profile to correct for expected correlation. `rehh` is another package used for selection scans, and for single populations uses extended haplotype homozygosity (EHH) and related statistics [@Gautier:2017]. These are popular and easy to interpret methods; however, they are not as effective for soft sweeps [@Liebert:2017]. Variations in local recombination rate are accounted for by implementing the cross-population EHH (XP-EHH) [@Sabeti:2007], which requires another population to compare with [@Vitti:2013]. All these packages identify evidence of selective sweeps in different ways, and each have merits. It would be advisable to use a combination of approaches to achieve the most accurate and valid results.
 
 # Conclusion
 
