+++ paper/paper.bib
@@ -1,17 +1,17 @@
 @book{Cutter:2019,
   	Author = {{Cutter}, A.D.},
   	Publisher = {Oxford University Press},
-  	Title = {{A Primer of Molecular Population Genetics}},
+  	Title = "{{A Primer of Molecular Population Genetics}}",
   	Year = 2019
 }
 @Article{Gautier:2017,
     author = {{Gautier}, M., and {Klassmann}, A. and {Vitalis}, R.},
-    title = {rehh2.0: a reimplementation of the R package rehh to detect positive selection from haplotype structure},
+    title = "{rehh2.0: a reimplementation of the R package rehh to detect positive selection from haplotype structure}",
     journal = {Molecular Ecology Resources},
     year = {2017},
     volume = {17},
     number = {1},
-    pages = {78-90},
+    pages = {78-90}
 }
 @article{Jacobs:2016,
   title="{Refining the Use of Linkage Disequilibrium as a Robust Signature of Selective Sweeps}",
@@ -32,6 +32,15 @@
   year={2001},
   doi={10.1038/ng1001-217}
 }
+@article{Kelly:1997,
+   author = {{Kelly}, J.K.},
+   title = "{A test of neutrality based on interlocus associations}",
+   journal = {Genetics},
+   volume = {146},
+   number = {3},
+   pages = {1197-1206},
+   year = {1997}
+}
 @article{Kim:2004,
   title="{Linkage Disequilibrium as a Signature of Selective Sweeps}",
   author={{Kim}, Y. and {Nielsen}, R.},
@@ -42,6 +51,16 @@
   year={2004},
   doi={10.1534/genetics.103.025387}
 }
+@article{Liebert:2017,
+   author = {{Liebert}, A., and {Lopez}, S. and {Jones}, B.L. and {Montalva}, N. and {Gerbault}, P. and {Lau}, W. and {Thomas}, M.G. and {Bradman}, N. and {Maniatis}, N. and {Swallow}, D.M.},
+   title = "{World-wide distributions of lactase persistence alleles and the complex effects of recombination and selection}",
+   journal = {Human Genetics},
+   volume = {136},
+   number = {11-12},
+   pages = {1445-1453},
+   DOI = {10.1007/s00439-017-1847-y},
+   year = {2017}
+}
 @article{Maynard:1974,
   title="{The hitch-hiking effect of a favourable gene}",
   author={{Maynard Smith}, J. and {Haigh}, J.},
@@ -49,20 +68,48 @@
   volume={23},
   number={1},
   pages={23--35},
-  year={1974},
+  year={1974}
+}
+@article{Nielsen:2005,
+   author = {{Nielsen}, R., and {Williamson}, S. and {Kim}, Y. and {Hubisz}, M.J. and {Clark}, A.G. and {Bustamante}, C.},
+   title = "{Genomic scans for selective sweeps using SNP data}",
+   journal = {Genome Research},
+   volume = {15},
+   number = {11},
+   pages = {1566-1575},
+   DOI = {10.1101/gr.4252305},
+   year = {2005}
 }
 @Article{Pfeifer:2014,
-    title = {PopGenome: An Efficient Swiss Army Knife for Population Genomic Analyses in R},
+    title = "{PopGenome: An Efficient Swiss Army Knife for Population Genomic Analyses in R}",
     journal = {Molecular Biology and Evolution},
     year = {2014},
     author = {{Pfeifer}, B., and {Wittelsbuerger}, U. and {Ramos-Onsins}, S.E. and {Lercher}, M.J.},
     volume = {31},
     pages = {1929-1936},
-    doi = {10.1093/molbev/msu136},
[…]
