### 2020-11-23 #2638 @chorscroft
@lpantano @marcrr thanks very much to both of you! I have expanded the final paragraph to talk more about the other options available to researchers for running selection scans in R. You were both right, this does read better and make the paper stronger. Thank you very much for your time! Please let me know if there is anything else you'd like me to do.
