### 2020-11-25 #2638 @lpantano
@marcrr, I see some items uncheck, can you let us know if they are ok now, or any thing else needs to be addressed:

Installation instructions, Automated tests, State of the Field, References

Probably the two last are ok according JOSS guidelines. The two first are required.

Thanks!
