### 2020-11-30 #2638 @chorscroft
@lpantano Yes, the version on CRAN is v0.2.0
