### 2020-12-01 0ea73d5
Added the doi for the Gautier paper
