+++ paper/paper.bib
@@ -11,7 +11,8 @@
     year = {2017},
     volume = {17},
     number = {1},
-    pages = {78-90}
+    pages = {78-90},
+    doi = {10.1111/1755-0998.12634}
 }
 @article{Jacobs:2016,
   title="{Refining the Use of Linkage Disequilibrium as a Robust Signature of Selective Sweeps}",
