### 2020-12-02 #2638 @chorscroft
@lpantano I have now updated the doi in the paper, sorry about that! I have also archived the software v0.2.0 on Zenodo  and the doi for this is: 10.5281/zenodo.4301636

Please let me know if you need me to do anything else, thanks for your time so far!
