### 2020-12-04 #2638 @kyleniemeyer
Hi @chorscroft, I'm the EIC on duty this week doing some final checks. It looks like a few citations may be missing DOIs: Kelly, Smith, Rozas, and Sabeti. Could you check if those articles do have citations, and add if so? There's no need to archive the repository again, since the software isn't changing.
