### 2020-12-04 #2638 @kyleniemeyer
@lpantano mostly it's just that whedon's checking of references isn't perfect—it can catch issues, for sure, but I still find it important to scan myself for missing DOIs.
