+++ paper/paper.bib
@@ -69,7 +69,8 @@
   volume={23},
   number={1},
   pages={23--35},
-  year={1974}
+  year={1974},
+  doi={10.1017/S0016672300014634}
 }
 @article{Nielsen:2005,
    author = {{Nielsen}, R., and {Williamson}, S. and {Kim}, Y. and {Hubisz}, M.J. and {Clark}, A.G. and {Bustamante}, C.},
@@ -106,7 +107,8 @@
    volume = {449},
    number = {7164},
    pages = {913-918},
-   year = {2007}
+   year = {2007},
+   doi = {10.1038/nature06250}
 }
 @incollection{Vitti:2013,
   author       = {{Vitti}, J.J., and {Grossman}, S.R. and {Sabeti}, P.C.},
