### 2020-12-05 #2638 @chorscroft
@kyleniemeyer Apologies for that, I have added DOIs for the Maynard Smith paper and the Sabeti paper. Unfortunately I can't seem to find DOIs for [Kelly](https://www.genetics.org/content/146/3/1197) or [Rozas](https://www.genetics.org/content/158/3/1147). Is there anything else I can do?
Thanks!
